Extract BuildListingTree.tsx
