+++ client/Library/Components/EncounterLibraryViewModel.tsx
@@ -1,6 +1,5 @@
 import _ = require("lodash");
 import * as React from "react";
-import { Listable } from "../../../common/Listable";
 import { EncounterCommander } from "../../Commands/EncounterCommander";
 import { Overlay } from "../../Components/Overlay";
 import { SavedCombatant, SavedEncounter } from "../../Encounter/SavedEncounter";
@@ -8,7 +7,7 @@ import { UpdateLegacySavedEncounter } from "../../Encounter/UpdateLegacySavedEnc
 import { EncounterLibrary } from "../EncounterLibrary";
 import { FilterCache } from "../FilterCache";
 import { Listing } from "../Listing";
-import { Folder } from "./Folder";
+import { BuildListingTree } from "./BuildListingTree";
 import { LibraryFilter } from "./LibraryFilter";
 import { ListingViewModel } from "./Listing";
 import { ListingButton } from "./ListingButton";
@@ -26,35 +25,6 @@ interface State {
     previewIconHovered: boolean;
     previewWindowHovered: boolean;
     previewPosition: { left: number; top: number; };
-}
-
-export function BuildListingTree<T extends Listable>(
-    buildListingComponent: (listing: Listing<T>) => JSX.Element,
-    listings: Listing<T>[]): JSX.Element[]
-{
-    const rootListingComponents = [];
-    const folders = {};
-    listings.forEach(listing => {
-        if (listing.Path == "") {
-            const component = buildListingComponent(listing);
-
-            rootListingComponents.push(component);
-        } else {
-            if (folders[listing.Path] == undefined) {
-                folders[listing.Path] = [];
-            }
-            folders[listing.Path].push(listing);
-        }
-    });
-
-    const folderComponents = _.map(folders, (listings: Listing<T>[], folderName: string) => {
-        const listingComponents = listings.map(buildListingComponent);
-        return <Folder key={folderName} name={folderName}>
-            {listingComponents}
-        </Folder>;
-    });
-
-    return folderComponents.concat(rootListingComponents);
 }
 
 export class EncounterLibraryViewModel extends React.Component<EncounterLibraryViewModelProps, State> {
